refactor(dependencies): reorganize and update dependencies across multiple crates
- Removed unused dependencies such as `application-utils`, `futures`, and `serde` from various `Cargo.toml` files to streamline the project and reduce bloat.
- Introduced `dependent-utils` in place of `database-utils` in relevant crates to enhance modularity and maintainability.
- Updated import statements in the `lib.rs` files to reflect the changes in dependencies, ensuring that the code remains functional and clean.
- This refactor improves the overall structure of the project by consolidating utility functions and reducing unnecessary dependencies.
