+++ crates/services/statistics/Cargo.toml
@@ -7,8 +7,8 @@ edition = "2024"
 async-graphql = { workspace = true }
 common-models = { path = "../../models/common" }
 database-models = { path = "../../models/database" }
-database-utils = { path = "../../utils/database" }
 dependent-models = { path = "../../models/dependent" }
+dependent-utils = { path = "../../utils/dependent" }
 enum-models = { path = "../../models/enum" }
 hashbag = { workspace = true }
 itertools = { workspace = true }
