+++ crates/utils/database/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-application-utils = { path = "../../utils/application" }
 async-graphql = { workspace = true }
 background-models = { path = "../../models/background" }
 chrono = { workspace = true }
@@ -13,7 +12,6 @@ common-utils = { path = "../../utils/common" }
 database-models = { path = "../../models/database" }
 dependent-models = { path = "../../models/dependent" }
 enum-models = { path = "../../models/enum" }
-futures = { workspace = true }
 itertools = { workspace = true }
 jwt-service = { path = "../../services/jwt" }
 media-models = { path = "../../models/media" }
@@ -22,7 +20,6 @@ markdown = { workspace = true }
 rust_decimal = { workspace = true }
 rust_decimal_macros = { workspace = true }
 sea-orm = { workspace = true }
-serde = { workspace = true }
 supporting-service = { path = "../../services/supporting" }
 tracing = { workspace = true }
 user-models = { path = "../../models/user" }
