+++ crates/utils/database/src/lib.rs
@@ -1,48 +1,31 @@
-use std::{collections::HashMap, sync::Arc};
+use std::sync::Arc;
 
-use application_utils::{get_podcast_episode_by_number, get_show_episode_by_numbers};
 use async_graphql::{Error, Result};
 use background_models::{ApplicationJob, HpApplicationJob, LpApplicationJob};
-use chrono::{Timelike, Utc};
-use common_models::{
-    BackendError, DailyUserActivityHourRecord, DailyUserActivityHourRecordEntity, EntityAssets,
-    IdAndNamedObject,
-};
+use chrono::Utc;
+use common_models::{BackendError, EntityAssets, IdAndNamedObject};
 use common_utils::ryot_log;
 use database_models::{
-    access_link, collection, collection_to_entity, daily_user_activity, metadata,
+    access_link, collection, collection_to_entity,
     prelude::{
-        AccessLink, Collection, CollectionToEntity, DailyUserActivity, Exercise, Metadata, Review,
-        Seen, User, UserMeasurement, UserToEntity, Workout, WorkoutTemplate,
+        AccessLink, Collection, CollectionToEntity, Review, Seen, User, UserToEntity, Workout,
+        WorkoutTemplate,
     },
-    review, seen, user, user_measurement, user_to_entity, workout,
+    review, seen, user, user_to_entity, workout,
 };
-use dependent_models::{
-    ApplicationCacheKeyDiscriminants, ExpireCacheKeyInput, UserWorkoutDetails,
-    UserWorkoutTemplateDetails,
-};
-use enum_models::{EntityLot, MediaLot, SeenState, UserLot, Visibility};
+use dependent_models::{UserWorkoutDetails, UserWorkoutTemplateDetails};
+use enum_models::{EntityLot, UserLot, Visibility};
 
-use futures::TryStreamExt;
 use itertools::Itertools;
 use jwt_service::{Claims, verify};
 use markdown::to_html as markdown_to_html;
-use media_models::{
-    AnimeSpecifics, AudioBookSpecifics, BookSpecifics, MangaSpecifics, MediaCollectionFilter,
-    MediaCollectionPresenceFilter, MovieSpecifics, MusicSpecifics, PodcastSpecifics, ReviewItem,
-    SeenAnimeExtraInformation, SeenMangaExtraInformation, SeenPodcastExtraInformation,
-    SeenShowExtraInformation, ShowSpecifics, VideoGameSpecifics, VisualNovelSpecifics,
-};
+use media_models::{MediaCollectionFilter, MediaCollectionPresenceFilter, ReviewItem};
 use migrations::AliasedCollectionToEntity;
-use rust_decimal::{Decimal, prelude::ToPrimitive};
 use rust_decimal_macros::dec;
 use sea_orm::{
-    ActiveModelTrait, ActiveValue, ColumnTrait, DatabaseConnection, EntityTrait, FromQueryResult,
-    Order, QueryFilter, QueryOrder, QuerySelect, QueryTrait, Select,
-    prelude::{Date, DateTimeUtc, Expr},
-    sea_query::{NullOrdering, PgFunc},
+    ActiveModelTrait, ActiveValue, ColumnTrait, DatabaseConnection, EntityTrait, QueryFilter,
+    QueryOrder, QuerySelect, QueryTrait, Select, prelude::Expr, sea_query::PgFunc,
 };
-use serde::{Deserialize, Serialize};
 use supporting_service::SupportingService;
 use user_models::UserReviewScale;
 use uuid::Uuid;
@@ -412,362 +395,6 @@ pub async fn item_reviews(
         })
         .collect();
     Ok(all_reviews)
-}
-
-pub async fn calculate_user_activities_and_summary(
-    user_id: &String,
-    ss: &Arc<SupportingService>,
-    calculate_from_beginning: bool,
-) -> Result<()> {
-    #[derive(Debug, Serialize, Deserialize, Clone, FromQueryResult)]
-    struct SeenItem {
-        seen_id: String,
-        show_extra_information: Option<SeenShowExtraInformation>,
-        podcast_extra_information: Option<SeenPodcastExtraInformation>,
-        anime_extra_information: Option<SeenAnimeExtraInformation>,
-        manga_extra_information: Option<SeenMangaExtraInformation>,
-        metadata_id: String,
-        finished_on: Option<Date>,
-        last_updated_on: DateTimeUtc,
-        metadata_lot: MediaLot,
-        audio_book_specifics: Option<AudioBookSpecifics>,
-        book_specifics: Option<BookSpecifics>,
-        movie_specifics: Option<MovieSpecifics>,
-        music_specifics: Option<MusicSpecifics>,
-        podcast_specifics: Option<PodcastSpecifics>,
-        show_specifics: Option<ShowSpecifics>,
-        video_game_specifics: Option<VideoGameSpecifics>,
-        manual_time_spent: Option<Decimal>,
-        visual_novel_specifics: Option<VisualNovelSpecifics>,
-        anime_specifics: Option<AnimeSpecifics>,
-        manga_specifics: Option<MangaSpecifics>,
-    }
-
-    let start_from = match calculate_from_beginning {
-        true => {
-            DailyUserActivity::delete_many()
-                .filter(daily_user_activity::Column::UserId.eq(user_id))
-                .exec(&ss.db)
-                .await?;
-            Date::default()
-        }
-        false => DailyUserActivity::find()
-            .filter(daily_user_activity::Column::UserId.eq(user_id))
-            .order_by_with_nulls(
-                daily_user_activity::Column::Date,
-                Order::Desc,
-                NullOrdering::Last,
-            )
-            .one(&ss.db)
-            .await?
-            .and_then(|i| i.date)
-            .unwrap_or_default(),
-    };
-    let mut activities = HashMap::new();
-
-    fn get_activity_count<'a>(
-        activities: &'a mut HashMap<Option<Date>, daily_user_activity::Model>,
-        user_id: &'a String,
-        date: Option<Date>,
-        entity_id: String,
-        entity_lot: EntityLot,
-        metadata_lot: Option<MediaLot>,
-        timestamp: DateTimeUtc,
-    ) -> &'a mut daily_user_activity::Model {
-        ryot_log!(debug, "Updating activity counts for id: {:?}", entity_id);
-        let existing = activities
-            .entry(date)
-            .or_insert(daily_user_activity::Model {
-                date,
-                user_id: user_id.to_owned(),
-                ..Default::default()
-            });
-        existing.entity_ids.push(entity_id.clone());
-        let hour = if timestamp.minute() < 30 {
-            timestamp.hour()
-        } else {
-            timestamp.hour() + 1
-        };
-        let maybe_idx = existing.hour_records.iter().position(|hr| hr.hour == hour);
-        if let Some(idx) = maybe_idx {
-            existing.hour_records.get_mut(idx).unwrap().entities.push(
-                DailyUserActivityHourRecordEntity {
-                    entity_id,
-                    entity_lot,
-                    metadata_lot,
-                },
-            );
-        } else {
-            existing.hour_records.push(DailyUserActivityHourRecord {
-                hour,
-                entities: vec![DailyUserActivityHourRecordEntity {
-                    entity_id,
-                    entity_lot,
-                    metadata_lot,
-                }],
-            });
-        }
-        existing
-    }
-
-    let mut seen_stream = Seen::find()
-        .filter(seen::Column::UserId.eq(user_id))
-        .filter(seen::Column::State.eq(SeenState::Completed))
-        .filter(seen::Column::LastUpdatedOn.gt(start_from))
-        .left_join(Metadata)
-        .select_only()
-        .column_as(seen::Column::Id, "seen_id")
-        .columns([
-            seen::Column::ShowExtraInformation,
-            seen::Column::PodcastExtraInformation,
-            seen::Column::AnimeExtraInformation,
-            seen::Column::MangaExtraInformation,
-            seen::Column::MetadataId,
-            seen::Column::FinishedOn,
-            seen::Column::LastUpdatedOn,
-            seen::Column::ManualTimeSpent,
-        ])
-        .column_as(metadata::Column::Lot, "metadata_lot")
-        .columns([
-            metadata::Column::AudioBookSpecifics,
-            metadata::Column::BookSpecifics,
-            metadata::Column::MovieSpecifics,
-            metadata::Column::MusicSpecifics,
-            metadata::Column::PodcastSpecifics,
-            metadata::Column::ShowSpecifics,
-            metadata::Column::VideoGameSpecifics,
-            metadata::Column::VisualNovelSpecifics,
-            metadata::Column::AnimeSpecifics,
-            metadata::Column::MangaSpecifics,
-        ])
-        .into_model::<SeenItem>()
-        .stream(&ss.db)
-        .await?;
-
-    while let Some(seen) = seen_stream.try_next().await? {
-        let activity = get_activity_count(
-            &mut activities,
-            user_id,
-            seen.finished_on,
-            seen.seen_id,
-            EntityLot::Metadata,
-            Some(seen.metadata_lot),
-            seen.last_updated_on,
-        );
-        if let (Some(show_seen), Some(show_extra)) =
-            (seen.show_specifics, seen.show_extra_information)
-        {
-            if let Some(runtime) =
-                get_show_episode_by_numbers(&show_seen, show_extra.season, show_extra.episode)
-                    .and_then(|(_, e)| e.runtime)
-            {
-                activity.show_duration += runtime;
-            }
-        } else if let (Some(podcast_seen), Some(podcast_extra)) =
-            (seen.podcast_specifics, seen.podcast_extra_information)
-        {
-            if let Some(runtime) =
-                get_podcast_episode_by_number(&podcast_seen, podcast_extra.episode)
-                    .and_then(|e| e.runtime)
-            {
-                activity.podcast_duration += runtime;
-            }
-        } else if let Some(audio_book_extra) = seen.audio_book_specifics {
-            if let Some(runtime) = audio_book_extra.runtime {
-                activity.audio_book_duration += runtime;
-            }
-        } else if let Some(movie_extra) = seen.movie_specifics {
-            if let Some(runtime) = movie_extra.runtime {
-                activity.movie_duration += runtime;
-            }
-        } else if let Some(music_extra) = seen.music_specifics {
-            if let Some(runtime) = music_extra.duration {
-                activity.music_duration += runtime / 60;
-            }
-        } else if let Some(book_extra) = seen.book_specifics {
-            if let Some(pages) = book_extra.pages {
-                activity.book_pages += pages;
-            }
-        } else if let Some(visual_novel_extra) = seen.visual_novel_specifics {
-            if let Some(runtime) = visual_novel_extra.length {
-                activity.visual_novel_duration += runtime;
-            }
-        } else if let Some(_video_game_extra) = seen.video_game_specifics {
-            if let Some(manual_time_spent) = seen.manual_time_spent {
-                activity.video_game_duration +=
-                    (manual_time_spent / dec!(60)).to_i32().unwrap_or_default();
-            }
-        }
-        match seen.metadata_lot {
-            MediaLot::Book => activity.book_count += 1,
-            MediaLot::Show => activity.show_count += 1,
-            MediaLot::Music => activity.music_count += 1,
-            MediaLot::Anime => activity.anime_count += 1,
-            MediaLot::Movie => activity.movie_count += 1,
-            MediaLot::Manga => activity.manga_count += 1,
-            MediaLot::Podcast => activity.podcast_count += 1,
-            MediaLot::VideoGame => activity.video_game_count += 1,
-            MediaLot::AudioBook => activity.audio_book_count += 1,
-            MediaLot::VisualNovel => activity.visual_novel_count += 1,
-        };
-    }
-
-    let exercises = Exercise::find().all(&ss.db).await.unwrap();
-    let user_exercises = UserToEntity::find()
-        .filter(user_to_entity::Column::UserId.eq(user_id))
-        .filter(user_to_entity::Column::ExerciseId.is_not_null())
-        .all(&ss.db)
-        .await?;
-    let mut workout_stream = Workout::find()
-        .filter(workout::Column::UserId.eq(user_id))
-        .filter(workout::Column::EndTime.gte(start_from))
-        .stream(&ss.db)
-        .await?;
-    while let Some(workout) = workout_stream.try_next().await? {
-        let date = workout.end_time.date_naive();
-        let activity = get_activity_count(
-            &mut activities,
-            user_id,
-            Some(date),
-            workout.id,
-            EntityLot::Workout,
-            None,
-            workout.start_time,
-        );
-        activity.workout_count += 1;
-        activity.workout_calories_burnt += workout
-            .calories_burnt
-            .unwrap_or_default()
-            .to_i32()
-            .unwrap_or_default();
-        activity.workout_duration += workout.duration / 60;
-        let workout_total = workout.summary.total.unwrap();
-        activity.workout_personal_bests += workout_total.personal_bests_achieved as i32;
-        activity.workout_weight += workout_total.weight.to_i32().unwrap_or_default();
-        activity.workout_reps += workout_total.reps.to_i32().unwrap_or_default();
-        activity.workout_distance += workout_total.distance.to_i32().unwrap_or_default();
-        activity.workout_rest_time += (workout_total.rest_time as i32) / 60;
-        for exercise in workout.information.exercises {
-            let db_ex = exercises.iter().find(|e| e.id == exercise.id).unwrap();
-            if user_exercises
-                .iter()
-                .find(|e| e.exercise_id == Some(db_ex.id.clone()))
-                .unwrap()
-                .exercise_extra_information
-                .as_ref()
-                .map(|d| d.settings.exclude_from_analytics)
-                .unwrap_or_default()
-            {
-                continue;
-            }
-            activity.workout_exercises.push(db_ex.name.clone());
-            activity.workout_equipments.extend(db_ex.equipment);
-            activity.workout_muscles.extend(db_ex.muscles.clone());
-        }
-    }
-
-    let mut measurement_stream = UserMeasurement::find()
-        .filter(user_measurement::Column::UserId.eq(user_id))
-        .filter(user_measurement::Column::Timestamp.gte(start_from))
-        .stream(&ss.db)
-        .await?;
-    while let Some(measurement) = measurement_stream.try_next().await? {
-        let date = measurement.timestamp.date_naive();
-        let activity = get_activity_count(
-            &mut activities,
-            user_id,
-            Some(date),
-            measurement.timestamp.to_string(),
-            EntityLot::UserMeasurement,
-            None,
-            measurement.timestamp,
-        );
-        activity.measurement_count += 1;
-    }
-
-    let mut review_stream = Review::find()
-        .filter(review::Column::UserId.eq(user_id))
-        .filter(review::Column::PostedOn.gte(start_from))
-        .stream(&ss.db)
-        .await?;
-    while let Some(review) = review_stream.try_next().await? {
-        let date = review.posted_on.date_naive();
-        let activity = get_activity_count(
-            &mut activities,
-            user_id,
-            Some(date),
-            review.id,
-            EntityLot::Review,
-            None,
-            review.posted_on,
-        );
-        match review.entity_lot {
-            EntityLot::Person => activity.person_review_count += 1,
-            EntityLot::Exercise => activity.exercise_review_count += 1,
-            EntityLot::Metadata => activity.metadata_review_count += 1,
-            EntityLot::Collection => activity.collection_review_count += 1,
-            EntityLot::MetadataGroup => activity.metadata_group_review_count += 1,
-            _ => {}
-        }
-    }
-
-    for (_, activity) in activities.iter_mut() {
-        DailyUserActivity::delete_many()
-            .filter(daily_user_activity::Column::UserId.eq(user_id))
-            .filter(match activity.date {
-                None => daily_user_activity::Column::Date.is_null(),
-                Some(date) => daily_user_activity::Column::Date.eq(date),
-            })
-            .exec(&ss.db)
-            .await?;
-        ryot_log!(debug, "Inserting activity = {:?}", activity.date);
-        let total_review_count = activity.metadata_review_count
-            + activity.collection_review_count
-            + activity.metadata_group_review_count
-            + activity.person_review_count
-            + activity.exercise_review_count;
-        let total_metadata_count = activity.movie_count
-            + activity.show_count
-            + activity.podcast_count
-            + activity.anime_count
-            + activity.manga_count
-            + activity.music_count
-            + activity.audio_book_count
-            + activity.book_count
-            + activity.video_game_count
-            + activity.visual_novel_count;
-        let total_count = total_metadata_count
-            + activity.measurement_count
-            + activity.workout_count
-            + total_review_count;
-        let total_duration = activity.workout_duration
-            + activity.audio_book_duration
-            + activity.podcast_duration
-            + activity.movie_duration
-            + activity.show_duration
-            + activity.music_duration
-            + activity.visual_novel_duration
-            + activity.video_game_duration;
-        activity.hour_records.sort_by_key(|hr| hr.hour);
-        let mut model: daily_user_activity::ActiveModel = activity.clone().into();
-        model.id = ActiveValue::NotSet;
-        model.total_review_count = ActiveValue::Set(total_review_count);
-        model.total_metadata_count = ActiveValue::Set(total_metadata_count);
-        model.total_count = ActiveValue::Set(total_count);
-        model.total_duration = ActiveValue::Set(total_duration);
-        model.insert(&ss.db).await.unwrap();
-    }
-
-    ss.cache_service
-        .expire_key(ExpireCacheKeyInput::BySanitizedKey {
-            user_id: Some(user_id.to_owned()),
-            key: ApplicationCacheKeyDiscriminants::UserAnalytics,
-        })
-        .await?;
-
-    ryot_log!(debug, "Expired cache key for user: {:?}", user_id);
-
-    Ok(())
 }
 
 pub async fn deploy_job_to_calculate_user_activities_and_summary(
